Update getting started to reflect @hapi/hapi use
This seems to be the new recommended approach, unless you use the npm alias functionality, but that seems out of scope for Getting Started and is instead an npm item. A more sweeping change across the site might be warranted, if one is coming then feel free to close in favor of that.